revert relocation for BenefitSurtax funciton
