+++ taxcalc/functions.py
@@ -16,6 +16,7 @@ import pandas as pd
 import math
 import numpy as np
 from .decorators import iterate_jit, jit
+import copy
 
 
 @iterate_jit(nopython=True)
@@ -1420,3 +1421,38 @@ def ExpandIncome(_fica_was, e02400, c02500, c00100, e00400):
                         non_taxable_ss_benefits +
                         employer_share_fica)
     return _expanded_income
+
+
+def BenefitSurtax(self):
+        """
+        BenefitSurtax function: ...
+        """
+        nobenefits_calc = copy.deepcopy(self)
+        if self.policy.ID_BenefitSurtax_crt != 1.:
+                    # hard code the reform
+            nobenefits_calc.policy.ID_Medical_HC = \
+                int(nobenefits_calc.policy.ID_BenefitSurtax_Switch[0])
+            nobenefits_calc.policy.ID_StateLocalTax_HC = \
+                int(nobenefits_calc.policy.ID_BenefitSurtax_Switch[1])
+            nobenefits_calc.policy.ID_RealEstate_HC = \
+                int(nobenefits_calc.policy.ID_BenefitSurtax_Switch[2])
+            nobenefits_calc.policy.ID_Casualty_HC = \
+                int(nobenefits_calc.policy.ID_BenefitSurtax_Switch[3])
+            nobenefits_calc.policy.ID_Miscellaneous_HC = \
+                int(nobenefits_calc.policy.ID_BenefitSurtax_Switch[4])
+            nobenefits_calc.policy.ID_InterestPaid_HC = \
+                int(nobenefits_calc.policy.ID_BenefitSurtax_Switch[5])
+            nobenefits_calc.policy.ID_Charity_HC = \
+                int(nobenefits_calc.policy.ID_BenefitSurtax_Switch[6])
+            nobenefits_calc.calc_one_year()
+            # pylint: disable=protected-access
+            tax_diff = np.where(
+                nobenefits_calc.records._iitax - self.records._iitax > 0.,
+                nobenefits_calc.records._iitax - self.records._iitax,
+                0.)
+            surtax_cap = nobenefits_calc.policy.ID_BenefitSurtax_crt *\
+                nobenefits_calc.records.c00100
+            _surtax = self.policy.ID_BenefitSurtax_trt *\
+                np.where(tax_diff > surtax_cap, tax_diff - surtax_cap, 0.)
+            self.records._iitax += _surtax
+            self.records._combined = self.records._iitax + self.records._fica
